requirements.txt
black==22.6.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pymongo==4.1.1
python-dotenv==0.20.0
tomli==2.0.1
Werkzeug==2.1.2
gunicorn
